pyproject.toml
[project]
name = "projekt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dash>=3.0.2",
    "loguru>=0.7.3",
    "pandas>=2.2.3",
    "plotly>=6.0.1",
    "psycopg2-binary>=2.9.10",
    "pyaml>=25.1.0",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.40",
]
